# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Montenegrina API
  version: 0.1.0
  description: Provider-independent Montenegrin conversational-AI control and data plane.
servers:
  - url: http://localhost:3001
security:
  - apiKey: []
  - sessionCookie: []
...
paths:
...
  /v1/organizations:
    get:
      operationId: listOrganizations
      tags: [Organizations]
      responses:
        '200': { $ref: '#/components/responses/OrganizationList' }
    post:
      operationId: createOrganization
      tags: [Organizations]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationRequest' }
      responses:
        '201': { $ref: '#/components/responses/Organization' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/organizations/{organizationId}:
    delete:
      operationId: deleteOrganization
      tags: [Organizations]
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { $ref: '#/components/responses/Deletion' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/organizations/{organizationId}/memberships:
    get:
      operationId: listMemberships
      tags: [Organizations]
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Memberships
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Membership' }
    post:
      operationId: createMembership
      tags: [Organizations]
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: '#/components/schemas/MembershipRole' }
      responses:
        '201':
          description: Membership created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Membership' }
  /v1/api-keys:
    post:
      operationId: createApiKey
      tags: [Organizations]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateApiKeyRequest' }
      responses:
        '201':
          description: API key shown once
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyCreated' }
...
  /v1/knowledge/bases/{knowledgeBaseId}/assignments:
    get:
      operationId: listKnowledgeBaseAssignments
      tags: [Knowledge]
      parameters:
        - name: knowledgeBaseId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Agent assignments
...
  /v1/knowledge/ingestion-jobs/{jobId}:
    get:
      operationId: getIngestionJob
      tags: [Knowledge]
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Ingestion job status
  /v1/knowledge/documents:
    get:
      operationId: listKnowledgeDocuments
      tags: [Knowledge]
      responses:
        '200': { $ref: '#/components/responses/DocumentList' }
    post:
      operationId: createKnowledgeDocument
      tags: [Knowledge]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [knowledgeBaseId, title, file]
              properties:
                knowledgeBaseId: { type: string, format: uuid }
                agentId: { type: string, format: uuid }
                title: { type: string, minLength: 1, maxLength: 300 }
                file: { type: string, format: binary }
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [agentId, title, text]
              properties:
                agentId: { type: string, format: uuid }
                title: { type: string, minLength: 1, maxLength: 300 }
                text: { type: string, minLength: 1 }
                sourceUrl: { type: string, format: uri }
      responses:
        '202': { $ref: '#/components/responses/Document' }
  /v1/knowledge/documents/{documentId}:
    delete:
      operationId: deleteKnowledgeDocument
      tags: [Knowledge]
      parameters:
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { $ref: '#/components/responses/Deletion' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: Authorization
    sessionCookie:
      type: apiKey
      in: cookie
      name: montenegrina_session
  parameters:
...
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: true }
...
    MembershipRole:
      type: string
      enum: [OWNER, ADMIN, DEVELOPER, VIEWER]
...
    Organization:
      type: object
      additionalProperties: false
      required: [id, name, slug, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        createdAt: { type: string, format: date-time }
    CreateOrganizationRequest:
      type: object
      additionalProperties: false
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, pattern: '^[a-z0-9][a-z0-9-]{1,62}$' }
    Membership:
      type: object
      additionalProperties: false
      required: [id, organizationId, userId, role]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/MembershipRole' }
    CreateApiKeyRequest:
      type: object
      additionalProperties: false
      required: [name, environment, permissions]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        environment: { type: string, enum: [development, staging, production] }
        permissions:
          type: array
          uniqueItems: true
          items: { type: string }
    ApiKeyCreated:
      type: object
      additionalProperties: false
      required: [id, prefix, key]
      properties:
        id: { type: string, format: uuid }
        prefix: { type: string }
        key: { type: string, description: Returned once and never stored verbatim. }
...
    Document:
      type: object
      additionalProperties: false
      required: [id, title, status, version, createdAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { type: string, enum: [UPLOADED, SCANNING, PROCESSING, READY, FAILED, QUARANTINED, DELETING] }
        version: { type: integer, minimum: 1 }
        errorCode: { type: string }
        createdAt: { type: string, format: date-time }
...
    Deletion:
      type: object
      additionalProperties: false
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [QUEUED, RUNNING, COMPLETED, FAILED] }
...
  responses:
    Error:
      description: Typed error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Organization:
      description: Organization
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Organization' }
    OrganizationList:
      description: Organizations
      content:
        application/json:
          schema:
            type: object
            required: [items]
            properties:
              items: { type: array, items: { $ref: '#/components/schemas/Organization' } }
...
    Document:
      description: Knowledge document
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Document' }
    DocumentList:
      description: Knowledge documents
      content:
        application/json:
          schema:
            type: object
            required: [items]
            properties:
              items: { type: array, items: { $ref: '#/components/schemas/Document' } }
...
    Deletion:
      description: Asynchronous deletion
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Deletion' }
...